Phase E (API ergonomics): rename simple → manual; doc pass; bump v0.14.0
Module rename:
- raptor::simple → raptor::manual (the adapter you build by hand with
  .route()/.footpath() calls; current name read as 'example code'). Updated
  every import site in test.rs, benches, proptest, dotgraph.

Docs updated to reflect the v0.14 builder API:
- lib.rs module docstring: query() builder example replaces the bare
  raptor() call.
- labels.rs ArrivalAndWalk doctest: query_with_label::<L>() builder.
- README: quick start, station-level queries, custom labels, per-leg timing,
  range queries, and 'Reusing a RaptorCache' all use .query()...run() chains
  with proper Tau/Duration newtype values.
- CHANGELOG: comprehensive v0.14 entry with new types, arithmetic, trait
  surface, builder usage, module rename, and migration guide.
- roadmap §3.1 (rRAPTOR): status updated to point at the new query builder.

Bumps Cargo.toml to v0.14.0.

58/58 tests + proptests + doctests green; clippy clean; cross-city bench
output unchanged within noise. The full ergonomics arc (A-E plus the
Tau-newtype migration) is now landed.
